Add files via upload
The **NIFTY50** is a benchmark stock market index in India, representing the weighted average of the 50 largest companies listed on the National Stock Exchange (NSE). It is widely regarded as an indicator of the performance of the Indian equity market and includes companies from various sectors. The NIFTY50 is used by investors to gauge the overall health of the market and to make informed investment decisions based on its movements.


### **Objective of the Code:**

The objective of this code is to analyze the historical performance of the NIFTY50 index by calculating the Compound Annual Growth Rate (CAGR) over various investment horizons (1 year, 2 years, etc.). The code aims to assess the likelihood of positive and negative returns over different periods by computing the percentage of instances with positive and negative CAGR values. This analysis helps investors understand the potential risks and returns associated with different investment durations in the NIFTY50 index.